Update to Babel 6
Disregarding the webpack build in this commit as Babel 6 now requires the babel-polyfill to handle generator functions.  Previously, we could just require the babel-runtime and it would transform the generator functions properly.  This is being tracked here:
https://phabricator.babeljs.io/T6676

Also, looks like the benchmark tests no longer run.  Not sure what that's about, but it generates invalid test suites now...